pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-trade"
version = "0.1.0"
description = "Coinbase Spot Trading Engine with limit entry and dynamic trailing exit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Trading System Team"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.28.0,<3.0",
    "aiohttp>=3.8.0,<4.0",
    "pytest>=7.0.0,<8.0",
    "pytest-asyncio>=0.21.0,<1.0",
    "aiohttp-session>=2.12.0,<3.0",
    "cryptography>=40.0.0,<42.0",
    "pyyaml>=6.0,<7.0",
    "loguru>=0.7.0,<1.0",
    "pydantic>=2.0.0,<3.0",
    "sqlcipher3-binary>=3.4.0,<4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0,<24.0",
    "ruff>=0.1.0,<1.0",
    "mypy>=1.0.0,<2.0",
    "isort>=5.12.0,<6.0",
    "pytest-cov>=4.0.0,<5.0",
    "pre-commit>=3.0.0,<4.0",
    "sphinx>=6.0.0,<8.0",
    "sphinx-rtd-theme>=1.2.0,<2.0",
]

monitoring = [
    "prometheus-client>=0.17.0,<1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/quant_trade"
Documentation = "https://quant-trade.readthedocs.io"
Repository = "https://github.com/yourusername/quant_trade.git"
Issues = "https://github.com/yourusername/quant_trade/issues"

[tool.setuptools]
packages = ["trading", "scripts"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function-call-in-default-argument
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F401", "F811"]

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "asyncio: marks tests as async",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["trading"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
